Add support for excluding directories.
It's now possible to stick an 'excludes' setting in the configuration
file. This setting defines either a single value, or an array of values,
with each value being a regular expression.

When loading the repository into memory, all directories that match any
of the exclusion filters are skipped.

I use it to treat directories starting with '_' special.